refactor(config): isolate implicit mention schema
