+++ src/config/zod-schema.channels-config.ts
@@ -3,6 +3,9 @@ import { z } from "zod";
 import type { ChannelsConfig } from "./types.channels.js";
 import { ChannelHeartbeatVisibilitySchema } from "./zod-schema.channels.js";
 import { ContextVisibilityModeSchema, GroupPolicySchema } from "./zod-schema.core.js";
+import { ChannelImplicitMentionsSchema } from "./zod-schema.implicit-mentions.js";
+
+export { ChannelImplicitMentionsSchema };
 
 const ChannelModelByChannelSchema = z
   .record(z.string(), z.record(z.string(), z.string()))
@@ -14,14 +17,6 @@ export const ChannelBotLoopProtectionSchema = z
     maxEventsPerWindow: z.number().int().positive().optional(),
     windowSeconds: z.number().int().positive().optional(),
     cooldownSeconds: z.number().int().positive().optional(),
-  })
-  .strict();
-
-export const ChannelImplicitMentionsSchema = z
-  .object({
-    replyToBot: z.boolean().optional(),
-    quotedBot: z.boolean().optional(),
-    threadParticipation: z.boolean().optional(),
   })
   .strict();
