+++ src/plugin-sdk/channel-config-schema.ts
@@ -26,5 +26,5 @@ export {
   requireAllowlistAllowFrom,
   requireOpenAllowFrom,
 } from "../config/zod-schema.core.js";
-export { ChannelImplicitMentionsSchema } from "../config/zod-schema.channels-config.js";
+export { ChannelImplicitMentionsSchema } from "../config/zod-schema.implicit-mentions.js";
 export { ToolPolicySchema } from "../config/zod-schema.agent-runtime.js";
